chore(v0.14): gate ruff and mypy in CI, with the lint policy pinned in-repo
- pyproject: [tool.ruff] pins a deliberate rule set (E4/E7/E9/F/W/I/B/C4/SIM/RUF,
  full-width punctuation ignored for Chinese docstrings) instead of "ALL" plus
  suppressions; [tool.mypy] sets a pragmatic, documented policy
- fixed everything both tools found: 145 ruff findings (100 auto, 45 triaged by
  hand) and 27 mypy errors - implicit params unions, a datasource variable that
  changed class per branch, a fetch import collision, a loop-closure binding in
  position.py, and four stale type: ignore comments
- CI: ruff check -> mypy src/aqlab -> pytest with the 85% coverage gate, still on
  the 3.10/3.11/3.12 matrix; dev extra gains ruff and mypy
- README (both languages): lint/type badges, highlights and roadmap updated

+++ src/aqlab/tables.py
@@ -28,7 +28,7 @@ def format_cell(value: Any) -> str:
         if value != value:  # NaN
             return "n/a"
         if abs(value) != float("inf") and abs(value - round(value)) < 1e-9 and abs(value) < 1e12:
-            return str(int(round(value)))
+            return str(round(value))
         return f"{value:.4f}".rstrip("0").rstrip(".")
     if isinstance(value, bool):
         return "true" if value else "false"
